Fix problems in text source data conversion and fix tests

+++ v2/googlecloud-to-neo4j/src/test/java/com/google/cloud/teleport/v2/neo4j/utils/DataCastingUtilsTest.java
@@ -501,7 +501,7 @@ public class DataCastingUtilsTest {
         ImmutableList.of(
             mapping("int64", PropertyType.Long),
             mapping("string", PropertyType.String),
-            mapping("double", PropertyType.BigDecimal),
+            mapping("double", PropertyType.Double),
             mapping("boolean", PropertyType.Boolean),
             mapping("localdate", PropertyType.Date),
             mapping("localtime", PropertyType.LocalTime),
